refactor: '전진하는 조건' 구현을 Car Class의 메소드로 변경
전진하는 조건은 Car Class에서만 사용되는 종속적인 함수이므로, Car Class의 메소드로 수정함
