+++ src/Car.js
@@ -1,7 +1,19 @@
+import { getRandomNumber } from "./util.js";
+
 export class Car {
   constructor(name) {
     this.name = name;
     this.position = 0;
+  }
+
+  canMoveForward() {
+    const MIN = 0;
+    const MAX = 9;
+    const BASE_POINT = 4;
+    
+    const randomNumber = getRandomNumber(MIN, MAX);
+  
+    return randomNumber >= BASE_POINT;
   }
 
   move() {
